+++ app/components/data-menu/services/data-layer-group-service.js
@@ -5,10 +5,11 @@
  * @memberof app
  * @name LayerGroup
  * @summary LayerGroup abstracts the notion of layers out of the app.
- * @description Only layergroups are approachable, from the outside world LayerGroup
- *              defines a group of layers which are loaded at initialization of the
- *              page. They can be toggled on/off and queried for data. Layergroup
- *              draws all its layers and returns data for all layers.
+ * @description Only layergroups are approachable, from the outside world
+ *              LayerGroup defines a group of layers which are loaded at
+ *              initialization of the page. They can be toggled on/off and
+ *              queried for data. Layergroup draws all its layers and returns
+ *              data for all layers.
  */
 angular.module('data-menu')
   .factory('DataLayerGroup', [
@@ -67,6 +68,15 @@ angular.module('data-menu')
         value: false,
         writable: true,
       });
+      Object.defineProperty(this, 'spatialBounds', {
+        value: {'east': 0, 'south': 0, 'north': 0, 'west': 0},
+        writable: true,
+      });
+      Object.defineProperty(this, 'temporalBounds', {
+        value: {'start': 0, 'end': 0},
+        writable: true,
+      });
+
 
       this.instantiateLayers(layerGroup.layers, layerGroup.temporal_resolution);
 
@@ -86,7 +96,9 @@ angular.module('data-menu')
         // map by the map-servie that go in mapLayers, layers that
         // are just used for data purposes are put in dataLayers
         // and layers that do both.
+
         angular.forEach(layers, function (layer) {
+
           if (layer.format === 'UTFGrid'
             || layer.format === 'Vector') {
             var nxtLayer = new NxtDataLayer(layer, tempRes);
@@ -95,6 +107,10 @@ angular.module('data-menu')
           }
           else if (layer.format === 'Store') {
             this._dataLayers.push(new NxtDataLayer(layer, tempRes));
+            if (layer.meta) {
+              this.temporalBounds = layer.meta.temporal_bounds;
+              this.spatialBounds = layer.meta.spatial_bounds;
+            }
           }
           else if (layer.format === 'TMS'
             || layer.format === 'WMS') {
@@ -117,8 +133,8 @@ angular.module('data-menu')
       },
 
       /**
-       * Returns true if the current layerGroup (i.e. "this") is active and false
-       * otherwise.
+       * Returns true if the current layerGroup (i.e. "this") is active and
+       * false otherwise.
        */
       isActive: function () {
         return this._active;
@@ -151,13 +167,13 @@ angular.module('data-menu')
       * @function
       * @memberOf app.LayerGroup.prototype
       * @description Returns a promise that notifies with data for every layer
-      *              of the layergroup that is appplicable (i.e: rain and several
-      *              vector layers). It resolves when all data is in.
-      * @param  {object} geom latLng object with lat and lng properties or a list of
-      *                       such objects.
-      * @return  {promise} notifies with data per layer and resolves with value true
-      *                    when layergroup was active, or false when layergroup was
-      *                    inactive.
+      *              of the layergroup that is appplicable (i.e: rain and
+      *              several vector layers). It resolves when all data is in.
+      * @param  {object} geom latLng object with lat and lng properties or a
+      *                  list of such objects.
+      * @return  {promise} notifies with data per layer and resolves with value
+      *                    true when layergroup was active, or false when
+      *                    layergroup was inactive.
       */
       getData: function (options) {
         var lgSlug = this.slug,