+++ app/components/graph/graph-directives.js
@@ -117,10 +117,11 @@ angular.module('lizard-nxt')
     };
 
     /**
-     * Calls updateGraph when data changes.
+     * Calls updateGraph when data is different than controller.data.
+     * NOTE: Controller data is set on precompile.
      */
     scope.$watch('data', function (n, o) {
-      if (n === o) { return true; }
+      if (n === graphCtrl.data) { return true; }
       graphUpdateHelper(true, false);
     });
 
@@ -139,9 +140,25 @@ angular.module('lizard-nxt')
       }
     });
 
-    scope.$on('$timelineZoomSuccess', function () {
-      graphUpdateHelper(true, true);
+    scope.$watch('temporal.start', function (n, o) {
+      if (n === o) { return true; }
+      graphUpdateHelper(false, true);
     });
+
+    scope.$watch('temporal.end', function (n, o) {
+      if (n === o) { return true; }
+      graphUpdateHelper(false, true);
+    });
+
+    scope.$watch('dimensions.height', function (n, o) {
+      if (!scope.dimensions
+        || scope.dimensions.height === graphCtrl.graph.dimensions.height) {
+        return true;
+      }
+      graphCtrl.graph.resize(scope.dimensions);
+      graphUpdateHelper(false, false);
+    });
+
   };
 
   /**