+++ app/components/omnibox/controllers/point-controller.js
@@ -16,14 +16,13 @@ angular.module('omnibox')
   '$scope',
   '$q',
   'LeafletService',
-  'TimeseriesService',
   'ClickFeedbackService',
   'UtilService',
   'MapService',
   'DataService',
   'State',
 
-  function ($scope, $q, LeafletService, TimeseriesService, ClickFeedbackService, UtilService, MapService, DataService, State) {
+  function ($scope, $q, LeafletService, ClickFeedbackService, UtilService, MapService, DataService, State) {
 
     var GRAPH_WIDTH = 600;
     $scope.box.content = {};
@@ -47,12 +46,6 @@ angular.module('omnibox')
       // Draw feedback when all promises resolved
       promise.then(drawFeedback, null, function (response) {
         if (response && response.data) {
-          // Apparently, we're dealing with the waterchain:
-          if (response.data.id && response.data.entity_name) {
-            getTimeSeriesForObject(
-              response.data.entity_name + '$' + response.data.id
-            );
-          }
           // If we deal with raster data....
           if (response.layerSlug === 'rain' && response.data && response.data.data !== null) {
             if ($scope.box.content[response.layerGroupSlug] === undefined) { return; }
@@ -151,44 +144,6 @@ angular.module('omnibox')
       }
     };
 
-    /**
-     * @function
-     * @memberOf app.pointCtrl
-     * @description gets timeseries from service
-     */
-    var getTimeSeriesForObject = function (objectId) {
-
-      TimeseriesService.getTimeseries(objectId, State.temporal)
-      .then(function (result) {
-
-        $scope.box.content.timeseries = $scope.box.content.timeseries || {};
-
-        if (result.length > 0) {
-          // We retrieved data for one-or-more timeseries, but do these actually
-          // contain measurements, or just metadata? We filter out the timeseries
-          // with too little measurements...
-          var filteredResult = [];
-          angular.forEach(result, function (value) {
-            if (value.events.length > 1) {
-              filteredResult.push(value);
-            }
-          });
-          if (filteredResult.length > 0) {
-            // IF we retrieve at least one timeseries with actual measurements,
-            // we put the retrieved data on the $scope:
-            $scope.box.content.timeseries.data = filteredResult;
-            $scope.box.content.timeseries.selectedTimeseries = filteredResult[0];
-          } else {
-            // ELSE, we delete the container object for timeseries:
-            delete $scope.box.content.timeseries;
-          }
-
-        } else {
-          delete $scope.box.content.timeseries;
-        }
-      });
-    };
-
     // Update when user clicked again
     $scope.$watch(State.toString('spatial.here'), function (n, o) {
       if (n === o) { return; }
@@ -213,7 +168,7 @@ angular.module('omnibox')
 
     // Clean up stuff when controller is destroyed
     $scope.$on('$destroy', function () {
-      DataService.reject();
+      DataService.reject('omnibox');
       $scope.box.content = {};
       ClickFeedbackService.emptyClickLayer(MapService);
     });