+++ app/components/time-ctx/time-ctx-controller.js
@@ -1,18 +1,20 @@
 angular.module('time-ctx')
   .controller("TimeCtxCtrl", ["$scope", "State", function ($scope, State) {
 
-  $scope.eventAggs = undefined;
-
-  $scope.state = State;
+  this.state = State;
 
   // dimensions are dependent on screen size.
   // this is calculated in directive.
-  $scope.dimensions = {};
+  this.dims = {};
+
+  this.content = {};
 
   // statistics (maybe get dynamically from event aggregation service?)
-  $scope.stats = ['max', 'min', 'mean', 'sum', 'median', 'count'];
+  this.stats = ['max', 'min', 'mean', 'sum', 'median', 'count'];
+
+  this.eventAggs = undefined;
 
   // default selection
-  $scope.selectedStat = $scope.stats[2];
+  this.selectedStat = this.selectedStat || this.stats[2];
 
 }]);