+++ app/components/timeline/timeline-directives.js
@@ -11,6 +11,7 @@
 angular.module('lizard-nxt')
   .directive('timeline',
              ["$q",
+              "$timeout",
               "RasterService",
               "UtilService",
               "Timeline",
@@ -18,6 +19,7 @@ angular.module('lizard-nxt')
               "DataService",
               "State",
               function ($q,
+                        $timeout,
                         RasterService,
                         UtilService,
                         Timeline,
@@ -29,7 +31,7 @@ angular.module('lizard-nxt')
 
     var timelineSetsTime = false,
 
-        showTimeline = false, // Is set by user clicking data label, when true
+        showTimeline = true, // Is set by user clicking data label, when true
                               // timeline is shown.
 
         dimensions = {
@@ -62,8 +64,10 @@ angular.module('lizard-nxt')
 
           timelineSetsTime = true;
           State.temporal.timelineMoving = true;
-          State.temporal.start = UtilService.getMinTime( scale.domain()[0].getTime() );
-          State.temporal.end   = UtilService.getMaxTime( scale.domain()[1].getTime() );
+          State.temporal.start = UtilService.getMinTime(
+            scale.domain()[0].getTime());
+          State.temporal.end   = UtilService.getMaxTime(
+             scale.domain()[1].getTime());
 
           State.temporal.aggWindow = UtilService.getAggWindow(
             State.temporal.start,
@@ -87,11 +91,10 @@ angular.module('lizard-nxt')
       zoomEndFn: function () {
         scope.$apply(function () {
           State.temporal.resolution = (
-            State.temporal.end - State.temporal.start) /  UtilService.getCurrentWidth();
+            State.temporal.end - State.temporal.start
+            ) /  UtilService.getCurrentWidth();
           getTimeLineData();
           State.temporal.timelineMoving = false;
-
-          scope.$broadcast("$timelineZoomSuccess");
         });
       },
 
@@ -108,7 +111,8 @@ angular.module('lizard-nxt')
       clickFn: function (event, scale, dimensions) {
         scope.$apply(function () {
           var timeClicked = +(scale.invert(
-            event.pageX - dimensions.padding.left - UtilService.TIMELINE_LEFT_MARGIN
+            event.pageX - dimensions.padding.left
+              - UtilService.TIMELINE_LEFT_MARGIN
           ));
           State.temporal.at = UtilService.roundTimestamp(
             timeClicked,
@@ -118,7 +122,8 @@ angular.module('lizard-nxt')
       },
     };
 
-    // shift timeline's SVG element using it's CSS - set here by JS too stop stuff becoming unsyncable
+    // shift timeline's SVG element using it's CSS - set here by JS too stop
+    // stuff becoming unsyncable
     angular.element("#timeline-svg-wrapper svg")[0].style.left
       = UtilService.TIMELINE_LEFT_MARGIN + "px";
 
@@ -165,6 +170,10 @@ angular.module('lizard-nxt')
         nEventTypes
       );
 
+      if (Timeline.onresize) {
+        Timeline.onresize(newDim);
+      }
+
     };
 
     /**
@@ -183,14 +192,19 @@ angular.module('lizard-nxt')
     var getTimelineLayers = function (layerGroups) {
       var timelineLayers = {events: {layers: [], slugs: []},
                             rain: undefined};
+
       angular.forEach(layerGroups, function (layergroup) {
         if (layergroup.isActive()) {
           angular.forEach(layergroup._dataLayers, function (layer) {
             if (layer.format === "Vector") {
               timelineLayers.events.layers.push(layer);
               timelineLayers.events.slugs.push(layer.slug);
-            } else if (layer.format === "Store" &&
-                       layer.slug === "rain") {
+            } else if (
+                // When time we do not want to draw rain in the timeline when
+                // context is time.
+                State.context !== 'time'
+                && layer.format === "Store"
+                && layer.slug === "rain") {
               timelineLayers.rain = layer;
             }
           });
@@ -310,11 +324,14 @@ angular.module('lizard-nxt')
             deferred: $q.defer()
           }
         }
-      ).then(function (response) {
-        if (response && response !== 'null') {
-          timeline.drawBars(response.data);
+      )
+      .then(
+        function (response) {
+          if (response && response !== 'null') {
+            timeline.drawBars(response.data);
+          }
         }
-      });
+      );
     };
 
     var timelineZoomHelper = function () {
@@ -339,15 +356,20 @@ angular.module('lizard-nxt')
 
     // END HELPER FUNCTIONS
 
-    element[0].style.height = 0;
-
     scope.timeline.toggleTimelineVisiblity = function () {
       showTimeline = !showTimeline;
-      if (!showTimeline) {
+      if (!showTimeline && State.context !== 'time') {
         element[0].style.height = 0;
       } else {
         updateTimelineHeight(scope.events.nEvents);
       }
+    };
+
+    scope.timeline.toggleTimelineVisiblity();
+
+    scope.timeline.toggleTimeCtx = function () {
+      scope.timeline.toggleTimelineVisiblity();
+      scope.transitionToContext(State.context === 'map' ? 'time' : 'map');
     };
 
     // WATCHES
@@ -377,15 +399,16 @@ angular.module('lizard-nxt')
       timelineZoomHelper();
     });
 
-    scope.$on("$timelineZoomSuccess", function () {
-      timelineZoomHelper();
-    });
-
     /**
      * Update aggWindow element when timeState.at changes.
      */
     scope.$watch(State.toString('temporal.at'), function (n, o) {
+      // update timeline when time-controller changes temporal.at state
       timeline.drawAggWindow(State.temporal.at, State.temporal.aggWindow);
+      // if temporal.playing don't get new data for each `temporal.at`
+      if (!State.temporal.playing) {
+        timelineZoomHelper();
+      }
     });
 
     /**
@@ -398,6 +421,13 @@ angular.module('lizard-nxt')
         State.temporal.aggWindow,
         false
       );
+    });
+
+    scope.$watch(State.toString('context'), function (n, o) {
+      if (n === o) { return; }
+      showTimeline = false; // It toggles
+      scope.timeline.toggleTimelineVisiblity();
+      getTimeLineData(); // It also removes data..
     });
 
     /**